improve FDC trace
git-svn-id: svn://svn.code.sf.net/p/steemsse/code/trunk@61 dd3fbd3b-ef5d-4b78-98e8-da144fa02e5b

+++ steemsse/steem/code/SSE/SSEShifter.cpp
@@ -1311,7 +1311,7 @@ BYTE TShifter::IORead(MEM_ADDRESS addr) {
       ASSERT( ST_TYPE==STE || !(xbios2&0xFF) );
       if(ST_TYPE==STE) 
 #endif
-        ior_byte=xbios2&0xFF;
+        ior_byte=(BYTE)xbios2&0xFF;
       break;
   
     case 0xff820f: // LINEWID
